+++ files/test.lua
@@ -21,12 +21,12 @@
 --at_test_y = 7500 -- rainforest dark
 --at_test_x = 12300 -- fungiforest
 --at_test_y = 2000 -- fungiforest
---at_test_x = -5317 -- hall of records
---at_test_y = 720 -- hall of records
+at_test_x = -5317 -- hall of records
+at_test_y = 720 -- hall of records
 --at_test_x = -5000 -- hall of records entrance
 --at_test_y = 700 -- hall of records entrace
-at_test_x = -5640 -- hall of records ghost
-at_test_y = 1024 -- hall of records ghost
+--at_test_x = -5640 -- hall of records ghost
+--at_test_y = 1024 -- hall of records ghost
 --at_test_x = -5433 -- hall of masters ghost
 --at_test_y = 1900 -- hall of masters ghost
 --at_test_x = 14334 --+ 35840 -- hall of masters ne gold